4.20.0

+++ app/src/main/java/com/zion830/threedollars/datasource/StoreDataSource.kt
@@ -48,12 +48,12 @@ interface StoreDataSource {
     ): Response<NewReviewResponse>
 
     suspend fun editReview(
-        reviewId: Int,
+        reviewId: Long,
         editReviewRequest: EditReviewRequest,
     ): Response<NewReviewResponse>
 
     suspend fun deleteReview(
-        reviewId: Int,
+        reviewId: Long,
     ): Response<BaseResponse<String>>
 
     fun getCategories(): Flow<Response<CategoriesResponse>>
