+++ app/src/main/java/com/zion830/threedollars/datasource/model/v2/response/my/MyReviewResponse.kt
@@ -23,7 +23,7 @@ data class ReviewDetail(
     @SerializedName("rating")
     val rating: Float = 0f,
     @SerializedName("reviewId")
-    val reviewId: Int = 0,
+    val reviewId: Long = 0L,
     @SerializedName("store")
     val store: StoreInfo = StoreInfo(),
     @SerializedName("storeName")
