+++ domain/src/main/java/com/threedollar/domain/home/data/store/EditStoreReviewModel.kt
@@ -4,7 +4,7 @@ data class EditStoreReviewModel(
     val contents: String,
     val createdAt: String,
     val rating: Int,
-    val reviewId: Int,
+    val reviewId: Long,
     val status: String,
     val storeId: Int,
     val updatedAt: String,
